requirements.txt
apache-airflow
azure-mgmt-hdinsight~=1.5.1
msrestazure~=0.6.3